requirements.txt
# AI Meeting Assistant - Dependencies

# Core dependencies
pytest>=7.4.0
pathlib

# LLM providers
openai>=1.0.0
anthropic>=0.18.0
zhipuai>=2.0.0  # 智谱 AI (GLM)

# ASR (Automatic Speech Recognition)
openai-whisper>=20230314
# faster-whisper>=0.10.0  # Optional: faster alternative

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0

# CLI & UI
rich>=13.0.0  # Terminal formatting
tqdm>=4.65.0  # Progress bars

# Note: ffmpeg must be installed separately
# Ubuntu/Debian: sudo apt-get install ffmpeg
# macOS: brew install ffmpeg
